test: move GuestRegionCollection specific tests to region.rs
Move some tests that are all about the invariants of
GuestRegionCollection constructors to region.rs, where they can be run
without the need for the backend-mmap feature (by instead using a mock
memory region).

While we're at it, fix these tests calling from_regions twice, but
from_arc_regions never.

Remove some test cases that are superfluous, because since the `regions`
field of `GuestRegionCollection` is private, all construction needs to
go through `from_regions`/`from_arc_regions`, and testing that wrappers
around these functions uphold the invariants of the wrapped functions is
not very useful.

test_memory and create_vec_with_regions were the same test, so
deduplicate while moving to region.rs.

Generally, most of these tests could be moved to region.rs, given
sufficient mocking of the memory region. I've somewhat arbitrarily drawn
the line at "only transfer tests where the mock only needs GuestAddress
and length", which roughly translates to "move tests where we are
testing the default implementations of the GuestMemory and
GuestMemoryRegion traits, which are not overwritten in the mmap-based
implementations". Of course, we could write a mock that implements
actual allocation of memory via std::alloc::alloc, but at that point
we'd be testing the mock more than actual vm-memory code (and start
loosing coverage of the mmap implementations).
